+++ Project.toml
@@ -2,7 +2,6 @@ name = "Bootstrap"
 uuid = "e28b5b4c-05e8-5b66-bc03-6f0c0a0a06e0"
 version = "2.0.1+"
 
-
 [deps]
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
@@ -13,7 +12,9 @@ StatsModels = "3eaba693-59b7-5ba5-a881-562e759f1c8d"
 
 [extras]
 GLM = "38e38edf-8417-5370-95a0-9cbb8c7f171a"
+RDatasets = "ce6b1742-4840-55fa-b093-852dadbb1d8b"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "GLM"]
+test = ["GLM", "RDatasets", "SafeTestsets", "Test"]
