+++ src/Bootstrap.jl
@@ -60,7 +60,6 @@ include("bootsampling.jl")
 include("get.jl")
 include("show.jl")
 include("confint.jl")
-include("datasets/Datasets.jl")
 include("deprecates.jl")
 
 end
